add in eslint, move files to src, and use a dist dir
