+++ .eslintignore
@@ -1,2 +1,2 @@
 npm node_modules
-build
+dist
