+++ package.json
@@ -6,21 +6,29 @@
 	"scripts": {
 		"dev": "rollup --config rollup.config.js -w",
 		"build": "NODE_ENV=production rollup --config rollup.config.js",
-		"version": "node version-bump.mjs && git add manifest.json versions.json"
+		"version": "node version-bump.mjs && git add manifest.json versions.json",
+		"lint": "eslint ."
 	},
 	"keywords": [],
 	"author": "",
 	"license": "MIT",
 	"devDependencies": {
 		"@rollup/plugin-commonjs": "^21.0.1",
+		"@rollup/plugin-eslint": "^8.0.1",
 		"@rollup/plugin-node-resolve": "^13.1.3",
 		"@types/node": "^16.11.6",
 		"@typescript-eslint/eslint-plugin": "^5.2.0",
 		"@typescript-eslint/parser": "^5.2.0",
+		"babel-plugin-root-import": "^6.6.0",
 		"builtin-modules": "^3.2.0",
 		"esbuild": "0.13.12",
+		"eslint": "^8.8.0",
+		"eslint-config-airbnb-base": "^15.0.0",
+		"eslint-config-airbnb-typescript": "^16.1.0",
 		"obsidian": "^0.12.17",
 		"rollup": "^2.66.1",
+		"rollup-plugin-copy": "^3.4.0",
+		"rollup-plugin-root-import": "^1.0.0",
 		"rollup-plugin-terser": "^7.0.2",
 		"rollup-plugin-typescript2": "^0.31.1",
 		"rollup-plugin-web-worker-loader": "^1.6.1",
