--- palette-modal-adapters/index.ts
@@ -1,3 +1,0 @@
-export * from './command-adapter';
-export * from './file-adapter';
-export * from './tag-adapter';
