+++ src/palette-modal-adapters/file-adapter.ts
@@ -1,16 +1,30 @@
-import { App, normalizePath } from "obsidian";
-import { Match } from "types";
-import { OrderedSet, PaletteMatch, SuggestModalAdapter, UnsafeAppInterface } from "utils";
+import { App, normalizePath } from 'obsidian';
+import {
+    OrderedSet, PaletteMatch, SuggestModalAdapter,
+} from 'src/utils';
+import { Match, UnsafeAppInterface } from 'src/types/types';
 
-export class BetterCommandPaletteFileAdapter extends SuggestModalAdapter {
+export default class BetterCommandPaletteFileAdapter extends SuggestModalAdapter {
+    titleText: string;
+
+    emptyStateText: string;
+
     // Unsafe interface
     app: UnsafeAppInterface;
 
     allItems: Match[];
+
     fileSearchPrefix: string;
 
-    constructor(app: App, prevItems: OrderedSet<Match>, recentAbovePinned: boolean, fileSearchPrefix: string) {
-        super(app, prevItems, recentAbovePinned)
+    constructor(
+        app: App,
+        prevItems: OrderedSet<Match>,
+        recentAbovePinned: boolean,
+        fileSearchPrefix: string,
+    ) {
+        super(app, prevItems, recentAbovePinned);
+        this.titleText = 'Better Command Palette: Files';
+        this.emptyStateText = 'No matching files.⌘+Enter to create the file.';
         this.fileSearchPrefix = fileSearchPrefix;
     }
 
@@ -23,16 +37,8 @@ export class BetterCommandPaletteFileAdapter extends SuggestModalAdapter {
     }
 
     cleanQuery(query: string): string {
-        const newQuery = query.replace(this.fileSearchPrefix, '')
+        const newQuery = query.replace(this.fileSearchPrefix, '');
         return newQuery;
-    }
-
-    getTitleText(): string {
-        return 'Better Command Palette: Files';
-    }
-
-    getEmptyStateText(): string {
-        return 'No matching files. ⌘+Enter to create the file.';
     }
 
     renderSuggestion(match: Match, el: HTMLElement): void {
@@ -42,9 +48,10 @@ export class BetterCommandPaletteFileAdapter extends SuggestModalAdapter {
         });
     }
 
-    async onChooseSuggestion(match: Match, event: MouseEvent | KeyboardEvent) {
+    async onChooseSuggestion(possbileMatch: Match, event: MouseEvent | KeyboardEvent) {
         let created = false;
         let file = null;
+        let match = possbileMatch;
 
         if (!match) {
             created = true;
@@ -59,15 +66,14 @@ export class BetterCommandPaletteFileAdapter extends SuggestModalAdapter {
         }
 
         this.getPrevItems().add(match);
-        const workspace = this.app.workspace;
+        const { workspace } = this.app;
 
         if (event.metaKey && !created) {
-            const newLeaf = workspace.createLeafBySplit(workspace.activeLeaf)
+            const newLeaf = workspace.createLeafBySplit(workspace.activeLeaf);
             newLeaf.openFile(file);
             workspace.setActiveLeaf(newLeaf);
         } else {
             workspace.activeLeaf.openFile(file);
         }
-    };
-
+    }
 }
