+++ src/palette-modal-adapters/index.ts
@@ -1,0 +1,3 @@
+export { default as BetterCommandPaletteCommandAdapter } from './command-adapter';
+export { default as BetterCommandPaletteFileAdapter } from './file-adapter';
+export { default as BetterCommandPaletteTagAdapter } from './tag-adapter';
