+++ src/palette.ts
@@ -1,34 +1,49 @@
-import { App, SuggestModal } from "obsidian";
-import { OrderedSet, PaletteMatch, SuggestModalAdapter, UnsafeSuggestModalInterface } from "./utils";
+import { App, SuggestModal } from 'obsidian';
+import {
+    OrderedSet, PaletteMatch, SuggestModalAdapter,
+} from 'src/utils';
+import { Match, UnsafeSuggestModalInterface } from 'src/types/types';
 import {
     BetterCommandPaletteCommandAdapter,
     BetterCommandPaletteFileAdapter,
     BetterCommandPaletteTagAdapter,
-} from "palette-modal-adapters";
-import { Match } from './types';
-import BetterCommandPalettePlugin from "main";
+} from 'src/palette-modal-adapters';
+import BetterCommandPalettePlugin from 'src/main';
 
-class BetterCommandPaletteModal extends SuggestModal<Match> implements UnsafeSuggestModalInterface  {
-    ACTION_TYPE_COMMAND: number = 1;
-    ACTION_TYPE_FILES: number = 2;
-    ACTION_TYPE_TAGS: number = 3;
+class BetterCommandPaletteModal extends SuggestModal<Match> implements UnsafeSuggestModalInterface {
+    ACTION_TYPE_COMMAND = 1;
+
+    ACTION_TYPE_FILES = 2;
+
+    ACTION_TYPE_TAGS = 3;
 
     // Unsafe interface
     chooser: UnsafeSuggestModalInterface['chooser'];
+
     updateSuggestions: UnsafeSuggestModalInterface['updateSuggestions'];
 
     actionType: number;
+
     fileSearchPrefix: string;
+
     tagSearchPrefix: string;
+
     suggestionsWorker: Worker;
+
     currentSuggestions: Match[];
+
     lastQuery: string;
+
     modalTitleEl: HTMLElement;
+
     initialInputValue: string;
 
     commandAdapter: BetterCommandPaletteCommandAdapter;
+
     fileAdapter: BetterCommandPaletteFileAdapter;
+
     tagAdapter: BetterCommandPaletteTagAdapter;
+
     currentAdapter: SuggestModalAdapter;
 
     constructor(
@@ -38,7 +53,7 @@ class BetterCommandPaletteModal extends SuggestModal<Match> implements UnsafeSug
         prevTags: OrderedSet<Match>,
         plugin: BetterCommandPalettePlugin,
         suggestionsWorker: Worker,
-        initialInputValue: string = '',
+        initialInputValue = '',
     ) {
         super(app);
 
@@ -49,7 +64,7 @@ class BetterCommandPaletteModal extends SuggestModal<Match> implements UnsafeSug
         this.initialInputValue = initialInputValue;
 
         // The only time the input will be empty will be when we are searching commands
-        this.setPlaceholder('Select a command')
+        this.setPlaceholder('Select a command');
 
         // Set up all of our different adapters
         this.commandAdapter = new BetterCommandPaletteCommandAdapter(
@@ -76,7 +91,7 @@ class BetterCommandPaletteModal extends SuggestModal<Match> implements UnsafeSug
 
         // Add our custom title element
         this.modalTitleEl = createEl('p', {
-            cls: 'better-command-palette-title'
+            cls: 'better-command-palette-title',
         });
 
         // Update our action type before adding in our title element so the text is correct
@@ -94,12 +109,12 @@ class BetterCommandPaletteModal extends SuggestModal<Match> implements UnsafeSug
             // Have to cast this to access `value`
             const el = event.target as HTMLInputElement;
 
-            if (plugin.settings.closeWithBackspace && el.value == '') {
-                this.close()
+            if (plugin.settings.closeWithBackspace && el.value === '') {
+                this.close();
                 // Stops the editor from using the backspace event
                 event.preventDefault();
             }
-        }
+        };
 
         this.scope.register([], 'Backspace', (event: KeyboardEvent) => {
             closeModal(event);
@@ -120,14 +135,15 @@ class BetterCommandPaletteModal extends SuggestModal<Match> implements UnsafeSug
         super.onOpen();
 
         // Add the initial value to the input
-        // TODO: Figure out if there is a way to bypass the first seach result flickering before this is set
+        // TODO: Figure out if there is a way to bypass the first seach
+        // result flickering before this is set
         // As far as I can tell onOpen resets the value of the input so this is the first place
         if (this.initialInputValue) {
             this.inputEl.value = this.initialInputValue;
         }
     }
 
-	updateActionType() : boolean {
+    updateActionType() : boolean {
         const text: string = this.inputEl.value;
         let type;
 
@@ -138,7 +154,7 @@ class BetterCommandPaletteModal extends SuggestModal<Match> implements UnsafeSug
             type = this.ACTION_TYPE_TAGS;
             this.currentAdapter = this.tagAdapter;
         } else {
-            type = this.ACTION_TYPE_COMMAND
+            type = this.ACTION_TYPE_COMMAND;
             this.currentAdapter = this.commandAdapter;
         }
 
@@ -159,21 +175,20 @@ class BetterCommandPaletteModal extends SuggestModal<Match> implements UnsafeSug
     }
 
     updateTitleText() {
-        this.modalTitleEl.setText(this.currentAdapter.getTitleText())
+        this.modalTitleEl.setText(this.currentAdapter.getTitleText());
     }
 
     updateEmptyStateText() {
         this.emptyStateText = this.currentAdapter.getEmptyStateText();
     }
 
-
-	getItems(): Match[] {
+    getItems(): Match[] {
         return this.currentAdapter.getSortedItems();
     }
 
     receivedSuggestions(msg : MessageEvent) {
-        const results = msg.data.slice(0, this.limit)
-        const matches = results.map((r : Record<string, string>) => new PaletteMatch(r.id, r.text))
+        const results = msg.data.slice(0, this.limit);
+        const matches = results.map((r : Record<string, string>) => new PaletteMatch(r.id, r.text));
         this.currentSuggestions = matches;
         this.updateSuggestions();
     }
@@ -186,25 +201,24 @@ class BetterCommandPaletteModal extends SuggestModal<Match> implements UnsafeSug
         });
     }
 
-	getSuggestions(query: string): Match[] {
+    getSuggestions(query: string): Match[] {
         // The action type might have changed
         this.updateActionType();
 
         const getNewSuggestions = query !== this.lastQuery;
         this.lastQuery = query;
-        query = query.trim()
-        query = this.currentAdapter.cleanQuery(query);
+        const fixedQuery = this.currentAdapter.cleanQuery(query.trim());
 
         if (getNewSuggestions) {
             // Load suggestions in another thread
-            this.getSuggestionsAsync(query);
+            this.getSuggestionsAsync(fixedQuery);
         }
 
         // For now return what we currently have. We'll populate results later if we need to
         return this.currentSuggestions;
     }
 
-	renderPrevItems(match: Match, el: HTMLElement, prevItems: OrderedSet<Match>) {
+    renderPrevItems(match: Match, el: HTMLElement, prevItems: OrderedSet<Match>) {
         if (prevItems.has(match)) {
             el.addClass('recent');
             el.createEl('span', {
@@ -214,12 +228,12 @@ class BetterCommandPaletteModal extends SuggestModal<Match> implements UnsafeSug
         }
     }
 
-	renderSuggestion(match: Match, el: HTMLElement) {
+    renderSuggestion(match: Match, el: HTMLElement) {
         this.renderPrevItems(match, el, this.currentAdapter.getPrevItems());
         this.currentAdapter.renderSuggestion(match, el);
     }
 
-	async onChooseSuggestion(item: Match, event: MouseEvent | KeyboardEvent) {
+    async onChooseSuggestion(item: Match, event: MouseEvent | KeyboardEvent) {
         this.currentAdapter.onChooseSuggestion(item, event);
     }
 }
