+++ src/types/worker.d.ts
@@ -1,4 +1,4 @@
-declare module "web-worker:*" {
+declare module 'web-worker:*' {
     const WorkerFactory: new (options: any) => Worker;
     export default WorkerFactory;
 }
