--- types.d.ts
@@ -1,8 +1,0 @@
-export interface Comparable {
-    value: () => string;
-}
-
-export interface Match extends Comparable {
-    text: string,
-    id: string,
-}
